+++ client/src/components/ListingsBase.vue
@@ -16,30 +16,7 @@ export default {
   },
   data () {
     return {
-      dishes: [],
-
-      // new dish details
-      newDishName: '',
-      newDishIngredients: [],
-      newDishDietaryRestrictions: [],
-      newDishPrice: '',
-      newDishAmount: 0
-    }
-  },
-  methods: {
-    async newDish () {
-      let response = await DishesService.newDish({
-        name: this.newDishName,
-        ingredients: this.newDishIngredients,
-        dietaryRestrictions: this.newDishDietaryRestrictions,
-        price: this.newDishPrice,
-        amount: this.newDishAmount
-      })
-      if (response.data.success) {
-        alert('Success!')
-      } else {
-        alert('Error: ' + response.data.errorMessage)
-      }
+      dishes: []
     }
   },
   async created () {
